# api-swagger.yaml
openapi: 3.0.0
info:
  description: Ecommerce API
  version: '1.0.0'
  title: Ecommerce API
  termsOfService: 'http://swagger.io/terms/'
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
servers:
  - url: 'http://localhost:4000'
...
paths:
...
  /cities:
    post:
      tags:
        - cities
      summary: Create city
      operationId: createCity
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CityBody'
      responses:
        default:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Create city success'
                  result:
                    $ref: '#/components/schemas/City'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '422':
          description: Unprocessable entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
...
  /cities/all:
    get:
      tags:
        - cities
      summary: Get all city
      operationId: getAllCity
      security:
        - BearerAuth: []
      responses:
        default:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Get all city success'
                  data:
                    type: object
                    properties:
                      cities:
                        type: array
                        items:
                          $ref: '#/components/schemas/City'
                      page:
                        type: number
                        example: 1
                      limit:
                        type: number
                        example: 1
                      total_page:
                        type: number
                        example: 1
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
...
  /paymenttypes:
    post:
      tags:
        - paymenttypes
      summary: Create payment type
      operationId: createPaymentType
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentTypeBody'
      responses:
        default:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Create payment type success'
                  data:
                    $ref: '#/components/schemas/PaymentType'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
  /paymenttypes/update/{id}:
    put:
      tags:
        - paymenttypes
      summary: Update payment type
      operationId: updatePaymentType
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentTypeBody'
      responses:
        default:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Update payment type success'
                  data:
                    $ref: '#/components/schemas/PaymentType'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
  /paymenttypes/delete/{id}:
    delete:
      tags:
        - paymenttypes
      summary: Delete payment type
      operationId: deletePaymentType
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        default:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Delete payment type success'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
  /paymenttypes/{id}:
    get:
      tags:
        - paymenttypes
      summary: Get detail payment type
      operationId: getDetailPaymentType
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        default:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Get detail payment type success'
                  data:
                    $ref: '#/components/schemas/PaymentType'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
  /paymenttypes/all:
    get:
      tags:
        - paymenttypes
      summary: Get all payment type
      operationId: getAllPaymentType
      security:
        - BearerAuth: []
      responses:
        default:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Get all payment type success'
                  data:
                    type: object
                    properties:
                      paymenttypes:
                        type: array
                        items:
                          $ref: '#/components/schemas/PaymentType'
                      page:
                        type: number
                        example: 1
                      limit:
                        type: number
                        example: 10
                      total_page:
                        type: number
                        example: 1
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
...
components:
  schemas:
    Unauthorized:
      type: object
      properties:
        message:
          type: string
          example: 'Unauthorized'
...
    ValidationError:
      type: object
      properties:
        message:
          type: string
          example: 'Validation error'
        errors:
          type: object
          properties:
            email:
              type: object
              properties:
                type:
                  type: string
                  example: 'field'
                value:
                  type: string
                  example: 'tranhoangthiendev@gmail.com'
                msg:
                  type: string
                  example: 'Email or password is incorrect'
                path:
                  type: string
                  example: 'email'
                location:
                  type: string
                  example: 'body'
...
    CityBody:
      type: object
      properties:
        name:
          type: string
          example: "HCM"
    City:
      type: object
      properties:
        _id:
          type: string
          format: MongoId
          example: "69218ec683ccdc34f38abfbd"
        user_id:
          type: string
          format: MongoId
          example: "69218ec683ccdc34f38abfbd"
        name:
          type: string
          example: "HCM"
        created_at:
          type: string
          format: ISO8601
          example: "2025-11-22T10:21:58.648Z"
        updated_at:
          type: string
          format: ISO8601
          example: "2025-11-22T10:45:46.131Z"
...
    PaymentType:
      type: object
      properties:
        _id:
          type: string
          format: MongoId
          example: "69218ec683ccdc34f38abfbd"
        user_id:
          type: string
          format: MongoId
          example: "69218ec683ccdc34f38abfbd"
        name:
          type: string
          example: "PAYMENT_LATER"
        type:
          type: string
          example: "PAYMENT_LATER"
        created_at:
          type: string
          format: ISO8601
          example: "2025-11-22T10:21:58.648Z"
        updated_at:
          type: string
          format: ISO8601
          example: "2025-11-22T10:45:46.131Z"
...
    PaymentTypeBody:
      type: object
      properties:
        name:
          type: string
          example: "PAYMENT_LATER"
        type:
          type: string
          example: "PAYMENT_LATER"
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT